Change: moved in_docker() to utils module
