+++ goodbot/utils.py
@@ -2,11 +2,28 @@
 """
 utils.py module. Contains utility functions.
 """
+import os
+
 from pathlib import Path
 from typing import List
 
 ALLOWED_INSTRUCTIONS_SUFFIX = (".yaml", ".txt", "")
 
+def in_docker() -> bool:
+    """Checks if code is currently running in a Docker container.
+
+    Checks if Docker is in control groups or if there is a `.dockerenv`
+    file at the filesystem's root directory.
+
+    Returns:
+        bool: Whether or not the code is running in a Docker container.
+    """
+    path = "/proc/self/cgroup"
+    return (
+        os.path.exists("/.dockerenv")
+        or os.path.isfile(path)
+        and any("docker" in line for line in open(path))
+    )
 
 def is_scene(directory: Path) -> bool:
     """Checks if a directory is a scene that contains instructions.
